+++ core/examples/luigi-sample-angular/src/assets/luigi-config/main.js
@@ -1,6 +1,7 @@
-import { navigation } from './navigation';
 import { auth } from './auth';
+import { navigation } from './navigation';
 import { routing } from './routing';
+import { settings } from './settings';
 
 Luigi.setConfig({
   auth,
